Taking steps to add domain pieces for yesornosite
